Moved PC_parts declarations and collision logic
I moved the PC_body and blaster into the player object itself that way the positioning is easier to deal with. I also moved the bump_engine.hit() into a for loop [in playerMovementHandler()] that is supposed to check if the player is hitting a tile that should be bounded, but its not working. I think Im going to create hard-coded bounds for each level because bump.js and tile-mapping doesn't seem to mesh very easily unless Im missing something.